network-error-logging/sends-report-on-subdomain-404.https.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that NEL reports are sent for failed subdomain requests</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='./support/fetch.sub.js'></script>
</head>
<body>
  <script>
    promise_test(async t => {
      await fetchNELSubdomainPolicyResource();
      await fetchNELMissingResource('www2');
      assert_true(await reportExists({
        url: getNELMissingResourceURL('www2'),
        type: "network-error",
        report: {
          uri: getNELMissingResourceURL('www2'),
          sampling_fraction: 1.0,
          status_code: 404,
          type: "http.error",
        },
      }));
    });
  </script>
</body>
</html>
